backend/requirements.txt
# ── Web framework ──────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0

# ── Base de données ────────────────────────────────────────────────
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9       # driver PostgreSQL

# ── Authentification ───────────────────────────────────────────────
python-jose[cryptography]==3.3.0   # JWT
passlib[bcrypt]==1.7.4             # hachage des mots de passe
python-multipart==0.0.9            # upload de fichiers

# ── Validation & configuration ─────────────────────────────────────
pydantic==2.7.1
pydantic-settings==2.2.1
email-validator==2.1.1
python-dotenv==1.0.1

# ── Tests ──────────────────────────────────────────────────────────
pytest==8.2.0
httpx==0.27.0                      # client HTTP pour les tests FastAPI
bcrypt==4.0.1